requirements.txt
bidict==0.23.1
deprecated==1.2.18
flask-cors==5.0.0
flask-limiter==3.11.0
limits==4.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy
ordered-set==4.1.0
packaging==24.2
pandas
pygments==2.19.1
python-dateutil==2.9.0.post0
graphviz==0.20.3
pytz==2024.2
rich==13.9.4
six==1.16.0
typing-extensions==4.12.2
wrapt==1.17.2
requests
# 生产服务器依赖
gunicorn==23.0.0
gevent==24.2.1
greenlet==3.0.3
zope.event==5.0
zope.interface==6.4